requirements.txt
django==4.2.6
pillow==10.1.0
psycopg2-binary==2.9.9
djangorestframework==3.14.0
Markdown==3.5
django-filter==23.3
drf-yasg==1.21.7